Remove old inmemory fixture
